build view model
