+++ MainCore/Commands/UI/Tabs/Villages/BuildNormalCommand.cs
@@ -3,7 +3,7 @@ using Humanizer;
 using MainCore.Common.Models;
 using MainCore.UI.Models.Input;
 
-namespace MainCore.Commands.UI.Build
+namespace MainCore.Commands.UI.Tabs.Villages
 {
     [RegisterSingleton<BuildNormalCommand>]
     public class BuildNormalCommand
@@ -11,7 +11,6 @@ namespace MainCore.Commands.UI.Build
         private readonly IDialogService _dialogService;
         private readonly IMediator _mediator;
 
-        private readonly ITaskManager _taskManager;
         private readonly IValidator<NormalBuildInput> _normalBuildInputValidator;
         private readonly GetBuildings _getBuildings;
 
@@ -19,20 +18,12 @@ namespace MainCore.Commands.UI.Build
         {
             _dialogService = dialogService;
             _mediator = mediator;
-            _taskManager = taskManager;
             _normalBuildInputValidator = normalBuildInputValidator;
             _getBuildings = getBuildings;
         }
 
         public async Task Execute(AccountId accountId, VillageId villageId, NormalBuildInput normalBuildInput, int location)
         {
-            var status = _taskManager.GetStatus(accountId);
-            if (status == StatusEnums.Online)
-            {
-                _dialogService.ShowMessageBox("Warning", "Please pause account before modifing building queue");
-                return;
-            }
-
             var result = await _normalBuildInputValidator.ValidateAsync(normalBuildInput);
             if (!result.IsValid)
             {
@@ -52,7 +43,7 @@ namespace MainCore.Commands.UI.Build
 
         private async Task NormalBuild(AccountId accountId, VillageId villageId, NormalBuildPlan plan)
         {
-            var buildings = _getBuildings.Execute(villageId);
+            var buildings = _getBuildings.Layout(villageId);
 
             var building = buildings.Find(x => x.Location == plan.Location);
