+++ MainCore/Commands/UI/Tabs/Villages/BuildResourceCommand.cs
@@ -1,0 +1,40 @@
+﻿using FluentValidation;
+using MainCore.Common.Models;
+using MainCore.UI.Models.Input;
+
+namespace MainCore.Commands.UI.Tabs.Villages
+{
+    [RegisterSingleton<BuildResourceCommand>]
+    public class BuildResourceCommand
+    {
+        private readonly IMediator _mediator;
+        private readonly IDialogService _dialogService;
+        private readonly IValidator<ResourceBuildInput> _resourceBuildInputValidator;
+
+        public BuildResourceCommand(IMediator mediator, IDialogService dialogService, IValidator<ResourceBuildInput> resourceBuildInputValidator)
+        {
+            _mediator = mediator;
+            _dialogService = dialogService;
+            _resourceBuildInputValidator = resourceBuildInputValidator;
+        }
+
+        public async Task Execute(AccountId accountId, VillageId villageId, ResourceBuildInput resourceBuildInput)
+        {
+            var result = await _resourceBuildInputValidator.ValidateAsync(resourceBuildInput);
+            if (!result.IsValid)
+            {
+                _dialogService.ShowMessageBox("Error", result.ToString());
+                return;
+            }
+
+            var (type, level) = resourceBuildInput.Get();
+            var plan = new ResourceBuildPlan()
+            {
+                Plan = type,
+                Level = level,
+            };
+            new AddJobCommand().ToBottom(villageId, plan);
+            await _mediator.Publish(new JobUpdated(accountId, villageId));
+        }
+    }
+}
